+++ src/no/hig/imt3591/id3/nodes/BetaAttributeNode.java
@@ -4,12 +4,12 @@ package no.hig.imt3591.id3.nodes;
  * A node that goes to the right on a test on a continuous attribute.
  */
 public class BetaAttributeNode extends AttributeNode {
-    public BetaAttributeNode(int value, DecisionNode next) {
+    public BetaAttributeNode(double value, DecisionNode next) {
         super(value, next);
     }
 
     @Override
-    public boolean canVisit(int visitor) {
+    public boolean canVisit(double visitor) {
         return visitor > super.getValue();
     }
 }
